docker_config_API/requirements.txt
langchain-google-genai
langchain-core
langchain
google-api-core
pandas
fastapi
uvicorn
regex
chromadb
sentence-transformers
tf-keras
torch --index-url https://download.pytorch.org/whl/cu124
torchvision --index-url https://download.pytorch.org/whl/cu124
torchaudio --index-url https://download.pytorch.org/whl/cu124